requirements.txt
Django==5.2.6
psycopg2-binary==2.9.10
python-dotenv==1.1.1
dj-database-url==3.0.1

# Flake8 и плагины
flake8==5.0.4
flake8-blind-except==0.2.1
flake8-broken-line==0.6.0
flake8-bugbear==22.9.23
flake8-builtins==2.0.0
flake8-class-attributes-order==0.1.3
flake8-cognitive-complexity==0.1.0
flake8-comprehensions==3.10.0
flake8-debugger==4.1.2
flake8-eradicate==1.4.0
flake8-functions==0.0.7
flake8-isort==5.0.0
flake8-mutable==1.2.0
flake8-print==5.0.0
flake8-pytest-style==1.6.0
flake8-quotes==3.3.1
flake8-string-format==0.3.0
flake8-variables-names==0.0.5
isort==5.12.0
black==23.12.0
psycopg[binary]==3.2.10
python-dotenv==1.1.1
